requirements.txt
django==1.8
djangorestframework==3.3.1
